Create Danone_project_README.md
This Python code is intended to extract financial data from the company Danone from a CSV file and to calculate financial ratios such as the return on investment (ROI) and the return on equity (ROE). To make these calculations we decided to use Danone's year-end data on 31/06/2022 (retrieved from their website).
The code starts by importing several Python libraries such as numpy, scipy, sympy and pandas. Then it loads the data from the CSV file using the panda’s library and displays the data on the screen.
Then it extracts the values from the corresponding columns for the closing price, net profit, total assets and total liabilities. It then calculates ROI using the basic formula ROI = net profit / invested capital and ROE using the basic formula ROE = net profit / equity. Finally, it displays the ROI and ROE results on the screen